vips needs pygobject with py3
Using the vips formula with py3 was slightly annoying, you had to install
as:

	brew install pygobject3 --with-python3
	brew install vips

Or `pygobject3` was only installed for python2.

This change adds `with-python3` to the `pygobject3` dependency in the
vips formula, so you can now just run:

	brew install vips

I suppose an alternative would be to add a `--with-python3` option to
vips.