vips: revision for libspng dependency
libvips 8.10 now supports libspng for PNG read. This is a new PNG load
library that aims to be faster, simpler and more secure than libpng.

You can read about this new feature here:

https://libvips.github.io/libvips/2020/06/18/What's-new-in-8.10.html

This patch adds libspng as a dependency of vips.rb.

update the licence line: libvips is lgpl-2.1+