add libcgif
cgif is a GIF write library. It's small, fast, portable, has no
dependencies, and has good test coverage.

It has been adopted by libvips 8.12 to replace the imagemagick-based writer we
used previously. Writing with cgif is typically 2x faster and needs 10x less
memory.

We're hoping to get this cgif formula into homebrew in advance of the release
of libvips 8.12:

https://www.libvips.org/2021/11/14/What's-new-in-8.12.html